requirements.txt
prompt_toolkit==3.0.48
tabulate==0.9.0
wcwidth==0.2.13
python-telegram-bot
